Implementation of tempSpectrogram for uppsala data
Adds a method called tempSpectrogram() that so far only can read the data from the uppsala file. It makes a 2d graph with the PCOL option but it can be changed to a SURF1 or whatever you feel like BUT it takes forever to draw the surfacegraphs.

Will start work on a way of doing the same with the other data files.