Some minor global changes and added CSV compatability
Added a new function that can split strings given a character to split at. Also changed the name of the function to DataView3D which also has inputs now that makes it possible to specify which years to view, if its a "uppsala file" or CSV and what timewindow of the days you want to get your data from.